requirements.txt
langchain
llama-index
openai
python-dotenv
html2text